Add graphviz printing of the module hierarchy
Implement a simple graphviz exporter that visit the module tree from a
root module and generate a pdf of the module hierarchy.
Using the filter function, one can omit modules which are not
interesting, suchs as fifos, etc.
This is a simple proof of concept, however it could be extended to
export more interesting visualization, such as buses.