Make the patch a bit more targeted, and optional
After reviewing the changes with xobs he pointed out that
on the ICE40 it may be problematic to consume the extra FF's.

Thus I have reformatted the patch and the uartbridge so
that you have to activate it by setting a flag of
`pipeline_addr=True` to the bridge.

I also reduced register usage by leaving the data path
as combinational. It does not drive the critical path,
just the address. This cuts the regsiter usage from
about 64 registers down to 32, which should help on
very small devices like the ICE40.