Use asynchronous transfers when loading kernel
I observed that loading a 427KB kernel took about 2 minutes, so I
switched to using 4Mbit serial and it still took about 2 minutes. This
change fixes that by load requests to continue while awaiting
confirmation that the CRC was valid and the load succeeded.

Prior to this change, with serial speed set to 4Mbits and CRC checking
enabled, a 427KB file would take 116 seconds to load.

With this change, it now takes 1.1 seconds.

I tested by randomly changing, deleting or inserting a byte of the
output stream.

I've attempted to keep the communication protocol backwards compatible.
So old versions of litex_term should still work with new BIOS code and
vice versa.