Add alternate timer mechanism to LiteX BIOS on POWER systems
All POWER systems have an architecturally defined decrementer and
timestamp counter.  On such systems, the LiteX timer peripheral
is redundant, but the LiteX BIOS assumes it is present on all SoCs.

Add a minimal interface from the LiteX BIOS code to the POWER
internal decrementer for SoCs without the LiteX timer peripheral.